# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GpsDevice {
  device_id: string;
  display_name: string;
  active_state: string;
}

export interface OdometerRange {
  device_id: string;
  from: string;
  to: string;
  begin_odometer_miles: number;
  end_odometer_miles: number;
  total_miles: number;
}

export interface DailyMileage {
  date: string;
  start_odometer_miles: number;
  end_odometer_miles: number;
  miles_driven: number;
}

export interface MileageSummary {
  device_id: string;
  display_name: string;
  from: string;
  to: string;
  total_miles: number;
  daily_logs: DailyMileage[];
}

export interface Project {
  id: number;
  project_number: string;
  created_at: string;
}

export interface ProjectInput {
  project_number: string;
}

export interface TeamLeader {
  id: number;
  name: string;
  created_at: string;
}

export interface TeamLeaderInput {
  name: string;
}

export interface LogEntry {
  id: number;
  device_id: string;
  device_name: string;
  start_date: string;
  end_date: string;
  begin_odometer: number;
  end_odometer: number;
  indirect_miles: number;
  personal_miles: number;
  direct_miles: number;
  total_miles: number;
  project_number: string;
  team_leader_name: string;
  created_at: string;
}

export interface LogEntryInput {
  device_id: string;
  device_name: string;
  start_date: string;
  end_date: string;
  begin_odometer: number;
  end_odometer: number;
  indirect_miles: number;
  personal_miles: number;
  direct_miles: number;
  project_number: string;
  team_leader_name: string;
}

export type GetMileageSummaryParams = {
device_id: string;
from: string;
to: string;
};

export type GetOdometerRangeParams = {
device_id: string;
from: string;
to: string;
};