# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /log-entries:
    get:
      operationId: listLogEntries
      tags: [logEntries]
      summary: List all mileage log entries
      responses:
        "200":
          description: List of log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
    post:
      operationId: createLogEntry
      tags: [logEntries]
      summary: Create a mileage log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogEntry"
  /log-entries/{id}:
    delete:
      operationId: deleteLogEntry
      tags: [logEntries]
      summary: Delete a log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    LogEntry:
      type: object
      required:
        - id
        - device_id
        - device_name
        - start_date
        - end_date
        - begin_odometer
        - end_odometer
        - indirect_miles
        - personal_miles
        - direct_miles
        - total_miles
        - project_number
        - team_leader_name
        - created_at
      properties:
        id:
          type: integer
        device_id:
          type: string
        device_name:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        begin_odometer:
          type: number
        end_odometer:
          type: number
        indirect_miles:
          type: number
        personal_miles:
          type: number
        direct_miles:
          type: number
        total_miles:
          type: number
        project_number:
          type: string
        team_leader_name:
          type: string
        created_at:
          type: string
    LogEntryInput:
      type: object
      required:
        - device_id
        - device_name
        - start_date
        - end_date
        - begin_odometer
        - end_odometer
        - indirect_miles
        - personal_miles
        - direct_miles
        - project_number
        - team_leader_name
      properties:
        device_id:
          type: string
        device_name:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        begin_odometer:
          type: number
        end_odometer:
          type: number
        indirect_miles:
          type: number
        personal_miles:
          type: number
        direct_miles:
          type: number
        project_number:
          type: string
        team_leader_name:
          type: string