# lib/api-zod/src/generated/types/logEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LogEntryInput {
  device_id: string;
  device_name: string;
  start_date: string;
  end_date: string;
  begin_odometer: number;
  end_odometer: number;
  indirect_miles: number;
  personal_miles: number;
  direct_miles: number;
  project_number: string;
  team_leader_name: string;
}